parses in xy slices of a parent block and runs them through a cache before 2d merging

+++ Parse.h
@@ -8,6 +8,7 @@
 #include "Block.h"
 #include "robin_hood.h"
 #include "Compression.h"
+//#include <omp.h>
 
 class Parse {
 public:
@@ -16,7 +17,7 @@ public:
   std::vector<std::vector<std::vector<Block>>> XBlocks;
   std::vector<std::vector<std::string>> MapInformation;
   int NumXBlocks, NumYBlocks, NumZBlocks;
-  int Cache2dHit, Cache2dMiss; 
+  float Cache2dHit, Cache2dMiss; 
 public:
   Parse();
   Parse(std::vector<std::string> Line);
